requirements.txt
awscli==1.11.74
boto==2.46.1
Django==1.11
django-cors-headers==2.0.2
django-cors-middleware==1.3.1
django-extensions==1.7.8
djangorestframework==3.6.2
ipython==3.1.0
setuptools==2.2
psycopg2==2.6
gunicorn==19.3.0
